+++ jwt/tagyou_claims.go
@@ -3,6 +3,6 @@ package jwt
 import jwt "github.com/golang-jwt/jwt/v5"
 
 type TagyouClaims struct {
-	UserId uint `json:"userId"`
+	UserId int64 `json:"userId"`
 	jwt.RegisteredClaims
 }
